Semi Approved Project Lacks Error handling and try catch blocks and memory management (using statement?)

+++ Application/MyRestaurant.Application.Query/Menus/MenuQueryHandler.cs
@@ -27,7 +27,7 @@ namespace MyRestaurant.Application.Query.Menus
                     menuForWeek.Add(new MenuQueryResult
                     {
                         Date = i,
-                        DayOfWeek = i.DayOfWeek,
+                        DayOfWeek = (int)i.DayOfWeek,
                         Meals = []
                     });
                 }
